fix(http): decompose dispatcher flow and unify input error mapping
